+++ sdmx-dl-provider-connectors/src/main/java/internal/sdmxdl/connectors/drivers/EurostatDriver.java
@@ -42,7 +42,7 @@ public final class EurostatDriver implements SdmxWebDriver {
                     .builder()
                     .name("ESTAT")
                     .alias("EUROSTAT")
-                    .description("Eurostat")
+                    .descriptionOf("Eurostat")
                     .driver(CONNECTORS_EUROSTAT)
                     .endpointOf("https://ec.europa.eu/eurostat/SDMX/diss-web/rest")
                     .websiteOf("https://ec.europa.eu/eurostat/data/database")